+++ tools/forecast/__main__.py
@@ -143,6 +143,7 @@ def _create_rainbow(args: argparse.Namespace) -> Rainbow:
                    chunk_size=args.chunk_size,
                    frequency=args.download_period,
                    token=args.token,
+                   layer=args.layer,
                    sensors=sensors)
 
 
@@ -281,6 +282,9 @@ if __name__ == "__main__":
     rainbow_parser = subparser.add_parser("rainbow", help="Rainbow")
     _add_sensors_params(rainbow_parser)
     rainbow_parser.add_argument("--token", type=str, required=True, help="Token to access Rainbow API")
+    rainbow_parser.add_argument("--layer", type=str, required=False, default="precip",
+                                choices=Rainbow.API_PRECIP_LAYERS,
+                                help=f"Rainbow API precip layer. Values of {Rainbow.API_PRECIP_LAYERS}")
     rainbow_parser.set_defaults(func=_create_rainbow)
 
     # WeatherCompany